lab-docker-deploy/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.8
pytest==8.3.4
requests==2.32.3
